Bug with collision as the gravity is so fast, but want to keep gravity accurate (9.8) in case we add more complex forces later. Gravity component added.

+++ Engine/Engine/Src/ECS/TileComponent.h
@@ -19,8 +19,8 @@ public:
 		
 		texture = Game::assets->GetTexture(id);
 
-		position.x = static_cast<int>(xPos);
-		position.y = static_cast<int>(yPos);
+		position.x = (float)(xPos);
+		position.y = (float)(yPos);
 
 		srcRect.x = srcX;
 		srcRect.y = srcY;
